Create EnvelopList file
